timer

+++ src/main/java/util/timer/HashedWheelTimer.java
@@ -1,6 +1,5 @@
 package util.timer;
 
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.Queue;
 import java.util.Set;
@@ -8,10 +7,13 @@ import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class HashedWheelTimer implements Timer{
 	
 	private volatile long startTime ; 
+	
+	private static final AtomicLong idG = new AtomicLong(1); 
 	
 	private Queue<HashWheelBucketNode> taskQue = new LinkedList<>() ; 
 	
@@ -34,7 +36,7 @@ public class HashedWheelTimer implements Timer{
 	private long tickInterval ; 
 	
 	private enum State{
-		INITIAL(0),STARTED(1),WAIT_START(2),SHUTDOWN(3);
+		INITIAL(0),STARTED(1),WAIT_START(2),SHUTTINGDOWN(3),SHUTEDOWN(4);
 		private int state ; 
 		private State(int state){
 			this.state = state ; 
@@ -49,12 +51,13 @@ public class HashedWheelTimer implements Timer{
 		for(int i = 0;i < bucketLen;i++){
 			this.buckets[i] = new HashWheelBucket() ; 
 		}
+		this.tickInterval = tickInterval ; 
 		this.workerThread = new Thread(new Worker()) ; 
 		this.maxPendingNum = maxPendingNum ; 
 	}
 	
 	@Override
-	public void shedule(Task task, TimeUnit unit, long time) {
+	public TimerCancelable shedule(Task task, TimeUnit unit, long time) {
 		if(task == null){
 			throw new NullPointerException("task not be null") ; 
 		}
@@ -67,26 +70,53 @@ public class HashedWheelTimer implements Timer{
 		}
 		start() ;
 		long deadTime = System.currentTimeMillis() + unit.toMillis(time) ;
-		long deadTick = (deadTime - startTime)/tickInterval ; 
+		long deadTick = (deadTime - startTime)/tickInterval ;
 		HashWheelBucketNode node = new HashWheelBucketNode(task,deadTick) ; 
 		this.taskQue.add(node) ; 
+		return node ; 
 	}
 
 	public void start() {
 		if(state.get() == State.INITIAL.state){
-			workerThread.start();
-			state.compareAndSet(State.INITIAL.state, State.WAIT_START.state) ; 
-		}else if(state.get() == State.SHUTDOWN.state){
+			synchronized (state) {
+				if(state.get() == State.INITIAL.state){					
+					workerThread.start();
+					state.compareAndSet(State.INITIAL.state, State.WAIT_START.state) ; 
+					System.out.println("started");
+				}
+			}
+		}
+		if(state.get() == State.SHUTTINGDOWN.state || state.get() == State.SHUTEDOWN.state){
 			throw new RejectedExecutionException("has stop") ;
 		}
 		if(state.get() == State.WAIT_START.state){
-			try {
-				startWaiter.await();
-			} catch (InterruptedException e) {
-				e.printStackTrace();
-			} 
+			synchronized (state) {
+				if(state.get() == State.WAIT_START.state){
+					try {
+						startWaiter.await();
+					} catch (InterruptedException e) {
+						e.printStackTrace();
+					} 
+				}
+			}
+			
 		}
 		
+	}
+	
+	public void shutdown(){
+		if(state.get() == State.SHUTTINGDOWN.state){
+			throw new RejectedExecutionException("has stop") ;
+		}
+		state.set(State.SHUTTINGDOWN.state);
+		while(state.get() != State.SHUTEDOWN.state){
+			try {
+				Thread.sleep(100);
+			} catch (InterruptedException e) {
+				e.printStackTrace();
+			}
+		}
+//		System.out.println("shuted down");
 	}
 
 	
@@ -99,13 +129,14 @@ public class HashedWheelTimer implements Timer{
 			startWaiter.countDown();  
 			while(state.get() == State.STARTED.state){
 				waitForNextTick() ; 
-				clearCancelTak() ; 
 				queTobucket() ;
+				clearCancelTak() ; 
 				HashWheelBucket bucket = buckets[(int) (tick%buckets.length)] ; 
+//				System.out.println(bucket);
 				bucket.doTasks() ;
 				tick++ ; 
 			}
-			
+			state.set(State.SHUTEDOWN.state);
 		}
 
 		private void waitForNextTick() {
@@ -123,13 +154,6 @@ public class HashedWheelTimer implements Timer{
 		}
 		
 	}
-	
-	@Override
-	public Set<Task> stop() {
-		this.state.set(State.SHUTDOWN.state);
-		Set<Task> set = new HashSet<>() ;
-		return set ; 
-	}
 
 	public void queTobucket() {
 		for(int i = 0;i < 10000;i++){
@@ -142,6 +166,7 @@ public class HashedWheelTimer implements Timer{
 			}
 			int idx = (int) (node.deadlineTick%buckets.length) ; 
 			buckets[idx].add(node) ; 
+			node.bucket = buckets[idx] ; 
 		}
 	}
 
@@ -151,24 +176,23 @@ public class HashedWheelTimer implements Timer{
 			if(node == null){
 				break ; 
 			}
-			node.remove() ; 
+			node.cancel = true ; 
+			node.remove() ;
 		}
 		
 	}
 
 	private class HashWheelBucket{
 
-        private HashWheelBucketNode head ;
+        private HashWheelBucketNode head ; 
         
         private HashWheelBucketNode tail ;
-		
+        
 		public void doTasks() {
 			HashWheelBucketNode node = this.head ;  
-			while(state.get() == State.STARTED.state){
-				if(node == null){
-					break; 
-				}
+			while(node != null && state.get() == State.STARTED.state){
 				if(node.deadlineTick > tick){
+					node = node.next ; 
 					continue ; 
 				}
 				node.task.run();
@@ -184,6 +208,7 @@ public class HashedWheelTimer implements Timer{
 			}else{
 				tail.next = node ; 
 				node.front = tail ; 
+				this.tail = node ; 
 			}
 		}
 		
@@ -229,10 +254,26 @@ public class HashedWheelTimer implements Timer{
 			result.bucket = null ; 
 			return result ; 
 		}
+
+		@Override
+		public String toString() {
+			StringBuilder sb = new StringBuilder("") ; 
+			HashWheelBucketNode p = this.head ;
+			while(p != null && p != tail){
+				sb.append(p+",") ; 
+				p = p.next ; 
+			}
+			sb.append(tail+",") ; 
+			return sb.toString() ; 
+		}
+		
+		
 		
 	}
 	
-	private class HashWheelBucketNode{
+	private class HashWheelBucketNode implements TimerCancelable{
+		
+		private long id ; 
 		
 		private long deadlineTick ;  
 		
@@ -251,6 +292,7 @@ public class HashedWheelTimer implements Timer{
 		public HashWheelBucketNode(Task task,long deadLineTick){
 			this.task = task ;
 			this.deadlineTick = deadLineTick ; 
+			this.id = idG.incrementAndGet() ; 
 		}
 		
 		public boolean hasCancel() {
@@ -262,10 +304,65 @@ public class HashedWheelTimer implements Timer{
 		}
 
 		public void remove() {
-			bucket.remove(this);
+			if(bucket != null){				
+				bucket.remove(this);
+			}
+		}
+
+		public void cancel(){
+			if(!hasCancel()){				
+				HashedWheelTimer.this.cancelQue.add(this) ; 
+				this.cancel = true ; 
+			}
+		}
+		
+		@Override
+		public String toString() {
+			return "id=" + id + "&deadlineTick=" + deadlineTick;
 		}
 		
 		
 	}
 	
+	public static void main(String[] args) throws InterruptedException{
+		HashedWheelTimer timer =  new HashedWheelTimer(10, 1000, 10000) ; 
+		TimerCancelable cancelable = null ; 
+		for(int i = 1;i <= 100;i++){
+			final int pos  = i ; 
+			TimerCancelable tmp = timer.shedule(new Task() {
+				
+				private int id = pos ; 
+				
+				@Override
+				public void run() {
+					System.out.println("run==============="+id);
+				}
+				
+			},TimeUnit.SECONDS, i);
+			if(i == 10){
+				cancelable = tmp ; 
+			}
+		}
+		
+		Thread.sleep(2000); 
+//		timer.shutdown();
+		cancelable.cancel(); 
+	}
+	
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
